Move standard plugins to entrypoints
