+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED] - YYYY-MM-DD
+
+### Changed
+- [#324](https://github.com/equinor/webviz-config/pull/324) - Now also `webviz-config`
+shipped plugins install themselves through the `webviz_config_plugins` entrypoints group.
+
 ## [0.2.0] - 2020-10-06
 
 ### Changed
