+++ webviz_config/generic_plugins/_banner_image.py
