+++ webviz_config/generic_plugins/_example_assets.py
