+++ webviz_config/generic_plugins/_table_plotter.py
